searcher/pyproject.toml
[tool.poetry]
name = "searcher"
version = "0.1.0"
description = ""
authors = ["Shawn Watanabe <nownabe@google.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
google-cloud-aiplatform = "^1.22.0"
tensorflow = "^2.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
